edi 파일 삭제 추가

+++ src/app/components/app-main/edi/edi-view/edi-view.component.ts
@@ -117,26 +117,30 @@ export class EdiViewComponent extends FComponentBase {
       this.setLoading(false);
     }
   }
-  async removeEDIFile(item: EDIUploadFileModel): Promise<void> {
-    this.setLoading();
-    const ret = await FExtensions.restTry(async() => await this.thisService.deleteEDIFile(item.thisPK),
-      e => this.fDialogService.error("delete", e));
-    this.setLoading(false);
-    if (ret.result) {
-      const index = this.uploadModel.fileList.indexOf(item);
-      if (index == this.uploadModel.fileList.length - 1) {
-        if (this.uploadModel.fileList.length - 1 > 0) {
-          this.activeIndex = this.uploadModel.fileList.length - 2;
-        } else {
-          this.activeIndex = 0;
+  async removeEDIFile(event: Event, item: EDIUploadFileModel): Promise<void> {
+    this.translateService.get(["common-desc.delete-sure", "common-desc.cancel", "common-desc.confirm"]).subscribe(x => {
+      this.confirmCall(event, x["common-desc.remove"], x["common-desc.delete-sure"], x["common-desc.cancel"], x["common-desc.confirm"], async() => {
+        this.setLoading();
+        const ret = await FExtensions.restTry(async() => await this.thisService.deleteEDIFile(item.thisPK),
+          e => this.fDialogService.error("delete", e));
+        this.setLoading(false);
+        if (ret.result) {
+          const index = this.uploadModel.fileList.indexOf(item);
+          if (index == this.uploadModel.fileList.length - 1) {
+            if (this.uploadModel.fileList.length - 1 > 0) {
+              this.activeIndex = this.uploadModel.fileList.length - 2;
+            } else {
+              this.activeIndex = 0;
+            }
+          }
+          if (index >= 0) {
+            this.uploadModel.fileList = [...this.uploadModel.fileList.filter(x => x.thisPK != item.thisPK)];
+          }
+          return;
         }
-      }
-      if (index >= 0) {
-        this.uploadModel.fileList = [...this.uploadModel.fileList.filter(x => x.thisPK != item.thisPK)];
-      }
-      return;
-    }
-    this.fDialogService.warn("delete", ret.msg);
+        this.fDialogService.warn("delete", ret.msg);
+      });
+    });
   }
 
   get acceptFiles(): string {
